# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Aiki — focus sessions, job tracking, streaks and badges.",
        "title": "Aiki API",
        "contact": {
            "name": "Aiki Support",
            "email": "support@aiki.app"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/jobs/recommended": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches jobs from SerpApi based on user profile (job title + experience level). Optional query param `location` (e.g. Austin, TX or United States) is passed to SerpApi as the job location filter. Returns cached results if fetched within 24 hours for the same location.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-search"
                ],
                "summary": "Get recommended jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job location filter (e.g. Seattle, WA) — forwarded to SerpApi Google Jobs",
                        "name": "location",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.JobSearchResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/jobs/recommended/{id}/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates a tracker row with status applied, links the Serp cache row, and returns apply_url (Google Jobs share link) for opening in an in-app browser or external browser.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-search"
                ],
                "summary": "Record an application and get the apply URL",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cache job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional notes",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.DirectApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.DirectApplyResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/jobs/recommended/{id}/save": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Saves a job from the recommended list into the user's job tracker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-search"
                ],
                "summary": "Save a recommended job to tracker",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cache job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DirectApplyRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                }
            }
        },
        "domain.DirectApplyResult": {
            "type": "object",
            "properties": {
                "apply_url": {
                    "type": "string"
                },
                "job": {
                    "$ref": "#/definitions/domain.Job"
                }
            }
        },
...
        "domain.Job": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_applied": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "platform": {
                    "description": "could be linkedin, website or indeed or others",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.JobSearchResult": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string"
                },
                "from_cache": {
                    "type": "boolean"
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SerpJobCache"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SerpJobCache": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "fetched_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "posted_at": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "saved_to_tracker": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "tracker_job_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by your access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}