# internal/database/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type BadgeDefinition struct {
	ID            int32            `json:"id"`
	Name          string           `json:"name"`
	Description   *string          `json:"description"`
	IconKey       *string          `json:"icon_key"`
	CriteriaType  string           `json:"criteria_type"`
	CriteriaValue int32            `json:"criteria_value"`
	CreatedAt     pgtype.Timestamp `json:"created_at"`
}

type DailyProgress struct {
	ID                int32       `json:"id"`
	UserID            int32       `json:"user_id"`
	Date              pgtype.Date `json:"date"`
	TotalFocusSeconds int32       `json:"total_focus_seconds"`
	SessionsCompleted int32       `json:"sessions_completed"`
}

type FocusSession struct {
	ID              int32            `json:"id"`
	UserID          int32            `json:"user_id"`
	DurationSeconds int32            `json:"duration_seconds"`
	ElapsedSeconds  int32            `json:"elapsed_seconds"`
	Status          string           `json:"status"`
	StartedAt       pgtype.Timestamp `json:"started_at"`
	EndedAt         pgtype.Timestamp `json:"ended_at"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
	UpdatedAt       pgtype.Timestamp `json:"updated_at"`
}

type Job struct {
	ID          int32            `json:"id"`
	UserID      int32            `json:"user_id"`
	Title       string           `json:"title"`
	CompanyName *string          `json:"company_name"`
	Notes       *string          `json:"notes"`
	Link        *string          `json:"link"`
	Location    *string          `json:"location"`
	Platform    *string          `json:"platform"`
	DateApplied pgtype.Timestamp `json:"date_applied"`
	Status      string           `json:"status"`
	CreatedAt   pgtype.Timestamp `json:"created_at"`
	UpdatedAt   pgtype.Timestamp `json:"updated_at"`
}

type Notification struct {
	ID        int32            `json:"id"`
	UserID    int32            `json:"user_id"`
	Type      string           `json:"type"`
	Title     string           `json:"title"`
	Message   string           `json:"message"`
	IsRead    bool             `json:"is_read"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type RefreshToken struct {
	ID        int32            `json:"id"`
	UserID    int32            `json:"user_id"`
	Token     string           `json:"token"`
	ExpiresAt pgtype.Timestamp `json:"expires_at"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type SerpJobCache struct {
	ID             int32            `json:"id"`
	UserID         int32            `json:"user_id"`
	ExternalID     string           `json:"external_id"`
	Title          string           `json:"title"`
	CompanyName    *string          `json:"company_name"`
	Location       *string          `json:"location"`
	Description    *string          `json:"description"`
	Link           *string          `json:"link"`
	Platform       *string          `json:"platform"`
	PostedAt       *string          `json:"posted_at"`
	Salary         *string          `json:"salary"`
	SavedToTracker bool             `json:"saved_to_tracker"`
	TrackerJobID   *int32           `json:"tracker_job_id"`
	FetchedAt      pgtype.Timestamp `json:"fetched_at"`
}

type Streak struct {
	ID              int32            `json:"id"`
	UserID          int32            `json:"user_id"`
	CurrentStreak   int32            `json:"current_streak"`
	LongestStreak   int32            `json:"longest_streak"`
	LastSessionDate pgtype.Date      `json:"last_session_date"`
	UpdatedAt       pgtype.Timestamp `json:"updated_at"`
}

type User struct {
	ID           int32            `json:"id"`
	FirstName    *string          `json:"first_name"`
	LastName     *string          `json:"last_name"`
	Email        string           `json:"email"`
	PhoneNumber  *string          `json:"phone_number"`
	PasswordHash *string          `json:"password_hash"`
	LinkedinID   *string          `json:"linkedin_id"`
	IsActive     *bool            `json:"is_active"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
	UpdatedAt    pgtype.Timestamp `json:"updated_at"`
}

type UserBadge struct {
	ID       int32            `json:"id"`
	UserID   int32            `json:"user_id"`
	BadgeID  int32            `json:"badge_id"`
	EarnedAt pgtype.Timestamp `json:"earned_at"`
}

type UserProfile struct {
	ID                int32            `json:"id"`
	UserID            int32            `json:"user_id"`
	Cv                []byte           `json:"cv"`
	FullName          *string          `json:"full_name"`
	CurrentJob        *string          `json:"current_job"`
	ExperienceLevel   *string          `json:"experience_level"`
	Goals             []string         `json:"goals"`
	JobSearchLocation string           `json:"job_search_location"`
	UpdatedAt         pgtype.Timestamp `json:"updated_at"`
}